# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: YieldForge API
  version: 0.1.0
  description: |-
    AI-powered Investment Copilot for Brazilian retail investors (FIIs + Fixed Income).

    **Money & rates:** monetary amounts are integer **centavos** (`*_centavos`) and
    rates are integer **basis points** (`*_bps`, 1 bp = 0.01%). They are always JSON
    integers, never floating-point.

    **Auth:** all routes are deny-by-default and require a valid session cookie except
    the public allowlist (`/healthz`, `/readyz`, `/version`, `POST /auth/register`,
    `POST /auth/login`). Identity is taken from the session, never from the request body.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /holdings/fii:
    get:
      tags: [holdings]
      summary: List the caller's FII holdings
      operationId: listFIIHoldings
      responses:
        "200":
          description: The caller's FII holdings (possibly empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FIIHoldingResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [holdings]
      summary: Create an FII holding
      operationId: createFIIHolding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FIIHoldingRequest"
      responses:
        "201":
          description: The created holding.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FIIHoldingResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /holdings/fii/{id}:
    put:
      tags: [holdings]
      summary: Update an owned FII holding
      operationId: updateFIIHolding
      parameters:
        - $ref: "#/components/parameters/HoldingID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FIIHoldingRequest"
      responses:
        "200":
          description: The updated holding.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FIIHoldingResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/HoldingNotFound"
    delete:
      tags: [holdings]
      summary: Delete an owned FII holding
      operationId: deleteFIIHolding
      parameters:
        - $ref: "#/components/parameters/HoldingID"
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/HoldingNotFound"
  /holdings/fixed-income:
    get:
      tags: [holdings]
      summary: List the caller's fixed-income holdings
      operationId: listFixedIncomeHoldings
      responses:
        "200":
          description: The caller's fixed-income holdings (possibly empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FixedIncomeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [holdings]
      summary: Create a fixed-income holding
      operationId: createFixedIncomeHolding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FixedIncomeRequest"
      responses:
        "201":
          description: The created holding.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FixedIncomeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /holdings/fixed-income/{id}:
    put:
      tags: [holdings]
      summary: Update an owned fixed-income holding
      operationId: updateFixedIncomeHolding
      parameters:
        - $ref: "#/components/parameters/HoldingID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FixedIncomeRequest"
      responses:
        "200":
          description: The updated holding.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FixedIncomeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/HoldingNotFound"
    delete:
      tags: [holdings]
      summary: Delete an owned fixed-income holding
      operationId: deleteFixedIncomeHolding
      parameters:
        - $ref: "#/components/parameters/HoldingID"
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/HoldingNotFound"
  /holdings/fixed-income/{id}/reconcile:
    post:
      tags: [holdings]
      summary: Confirm interest and/or report a contribution for an owned fixed-income holding
      description: |-
        SPEC-110 FR-1103. Both amounts are additive (never a replacement): invested_amount_centavos
        += confirmed_interest_centavos + contribution_centavos; total_contributed_centavos +=
        contribution_centavos only (interest never touches it); last_reconciled_at resets to now.
        contribution_centavos may be 0 (pure interest confirmation). Distinct from PUT, which is a
        plain metadata/correction edit and never touches total_contributed_centavos.
      operationId: reconcileFixedIncomeHolding
      parameters:
        - $ref: "#/components/parameters/HoldingID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FixedIncomeReconcileRequest"
      responses:
        "200":
          description: The reconciled holding.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FixedIncomeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/HoldingNotFound"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: yf_session
      description: |-
        Server-side session token in an HttpOnly cookie, issued by `POST /auth/login`.
        The cookie name is configurable via `AUTH_COOKIE_NAME` (default `yf_session`).
  parameters:
    HoldingID:
      name: id
      in: path
      required: true
      description: The holding's UUID.
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: The request body or a field is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: invalid request body
    Unauthorized:
      description: Authentication is required or the session is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: authentication required
    HoldingNotFound:
      description: No such holding owned by the caller.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: holding not found
...
  schemas:
    Error:
      type: object
      description: The generic error envelope used by all 4xx/5xx responses.
      properties:
        error:
          type: string
      required: [error]
...
    FIIHoldingRequest:
      type: object
      properties:
        ticker:
          type: string
          description: The FII ticker (e.g. `HGLG11`).
          example: HGLG11
        quantity:
          type: integer
          minimum: 1
          description: Number of cotas (positive whole number).
        average_price_centavos:
          type: integer
          format: int64
          minimum: 0
          description: Average cost per cota, in centavos (integer, never a float).
      required: [ticker, quantity, average_price_centavos]
    FIIHoldingResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ticker:
          type: string
        quantity:
          type: integer
        average_price_centavos:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, ticker, quantity, average_price_centavos, created_at, updated_at]
...
    FixedIncomeRequest:
      type: object
      properties:
        name:
          type: string
        institution:
          type: string
        invested_amount_centavos:
          type: integer
          format: int64
          minimum: 1
          description: Invested amount in centavos (integer, never a float).
        annual_rate_bps:
          type: integer
          minimum: 0
          description: |-
            Meaning depends on indexer_type (SPEC-109): the flat annual rate for
            prefixado; the percentage of CDI in bps-of-percent for cdi_percentual
            (12000 = 120%); the spread over IPCA in bps for ipca_spread (580 = +5.80%).
        indexer_type:
          type: string
          enum: [prefixado, cdi_percentual, ipca_spread]
          default: prefixado
          description: "\"\" / omitted defaults to prefixado (SPEC-109 BR-1093)."
        maturity_date:
          type: [string, "null"]
          format: date
          description: Maturity as `YYYY-MM-DD`; null/omitted for daily-liquidity holdings.
        liquidity_type:
          type: string
          enum: [daily, at_maturity]
      required: [name, institution, invested_amount_centavos, annual_rate_bps, liquidity_type]
    FixedIncomeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        institution:
          type: string
        invested_amount_centavos:
          type: integer
          format: int64
          description: |-
            The current principal/balance basis for accrual (SPEC-110): grows via both
            contributions and confirmed interest. No longer the cost basis by itself —
            see total_contributed_centavos.
        total_contributed_centavos:
          type: integer
          format: int64
          description: |-
            New in SPEC-110 (FR-1101): the lifetime sum of money the user has contributed —
            the true cost basis for growth. Changes only via an explicit contribution
            (creation or POST .../reconcile), never via interest.
        annual_rate_bps:
          type: integer
          description: The raw stored value — see FixedIncomeRequest.annual_rate_bps.
        indexer_type:
          type: string
          enum: [prefixado, cdi_percentual, ipca_spread]
        effective_annual_rate_bps:
          type: integer
          description: |-
            Computed, never persisted (SPEC-109): the resolved current annual rate —
            equal to annual_rate_bps for prefixado; resolved against the latest
            SELIC/CDI/IPCA (GET /market/indicators) for cdi_percentual/ipca_spread.
        estimated_interest_centavos:
          type: integer
          format: int64
          description: |-
            Computed, never persisted (SPEC-110 FR-1103): the simple-interest estimate
            accrued since last_reconciled_at — the pre-fill hint for POST .../reconcile.
        reconciliation_due:
          type: boolean
          description: |-
            Computed, never persisted (SPEC-110 FR-1105): true when last_reconciled_at's
            calendar month is strictly before the current one.
        maturity_date:
          type: [string, "null"]
          format: date
        liquidity_type:
          type: string
          enum: [daily, at_maturity]
        last_reconciled_at:
          type: string
          format: date-time
          description: |-
            New in SPEC-110: the accrual clock. Starts at created_at; resets on every
            reconciliation, and on any edit that changes invested_amount_centavos (FR-1102).
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, institution, invested_amount_centavos, total_contributed_centavos, annual_rate_bps, liquidity_type, last_reconciled_at, created_at, updated_at]
    FixedIncomeReconcileRequest:
      type: object
      description: SPEC-110 FR-1103. Both amounts are additive; contribution_centavos may be 0.
      properties:
        confirmed_interest_centavos:
          type: integer
          format: int64
          minimum: 0
          description: The confirmed (or adjusted) interest since last_reconciled_at.
        contribution_centavos:
          type: integer
          format: int64
          minimum: 0
          description: A new contribution — 0 for a pure interest confirmation.
      required: [confirmed_interest_centavos, contribution_centavos]
...
security:
  - cookieAuth: []